Move slideshow from utils to imutils
The slideshow is used to change the current image and relies heavily on
the api. It is therefore not a simple utility but more related to the
image utilities.

The further changes in api.__init__ and plugins.print are related to
mypy complaints.
