+++ vimiv/imutils/slideshow.py
